+++ crates/cargo-msrv-cli/src/values/fallback_release_source.rs
@@ -2,6 +2,7 @@ use crate::values::{CliValue, CliValues};
 use cargo_msrv_context::types::FallbackReleaseSource;
 
 pub const VALUES: CliValues<FallbackReleaseSource> = CliValues::new(&[
+    #[cfg(feature = "rust-releases-changelog-source")]
     CliValue::new("rust-changelog", FallbackReleaseSource::RustChangelog),
     #[cfg(feature = "rust-releases-github-source")]
     CliValue::new("github", FallbackReleaseSource::GitHub),