+++ crates/cargo-msrv-context/src/types/mod.rs
@@ -3,6 +3,11 @@
 //! These are part of the user interface, but are also used by the program itself, which is why
 //! they are defined here, and not by the `cargo-msrv-cli` crate.
 
+#[cfg(any(
+    feature = "rust-releases-changelog-source",
+    feature = "rust-releases-github-source",
+    feature = "rust-releases-dist-source"
+))]
 pub mod bundled_fallback;
 pub mod edition;
 pub mod list_msrv_variant;
@@ -11,6 +16,11 @@ pub mod output_format;
 pub mod release_source;
 pub mod tracing_target_option;
 
+#[cfg(any(
+    feature = "rust-releases-changelog-source",
+    feature = "rust-releases-github-source",
+    feature = "rust-releases-dist-source"
+))]
 pub use bundled_fallback::{
     BundledFallback, BundledMaxAge, FallbackReleaseSource, ParseBundledMaxAgeError,
 };